# public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel and Swagger",
        "description": "Getting started with Laravel and Swagger",
        "termsOfService": "",
        "contact": {
            "email": "name@example.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/auth/signup": {
            "post": {
                "description": "Return Successful Message",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Your email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Your password",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "Your password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "description": "Return access_token",
                "parameters": [
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "application/json",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "description": "X-XMLHttpRequest",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "remember_me",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "Auto Login",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "examples": {
                            "application/json": {
                                "access_token": "string",
                                "token_type": "Bearer",
                                "expires_at": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "get": {
                "description": "Return Message",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
...
}